setup-kit/telegram-bot/requirements.txt
python-telegram-bot[ext]==21.6
openai>=1.50.0
websockets>=12.0
# OpenSpace itself is installed from the bundled source dir (../../.. in Docker)
